src/requirements.txt
Homie4
requests